Fix Bugbot findings: BE64 unpack and menu/outline cleanup

+++ Common/source/langhash.c
@@ -740,14 +740,14 @@ static void diskvalue_to_value_v7(const tydiskvaluedata_v7 *disk, tyvaluerecord 
 		case directionvaluetype:
 			val->data.dirvalue = disk_to_host_int32(disk->dirvalue);
 			break;
-		case datevaluetype:
-			val->data.datevalue = (unsigned long) disk_to_host_int64((int64_t) disk->datevalue);
+	case datevaluetype:
+			val->data.datevalue = disk_to_host_int64((int64_t) disk->datevalue);
 			break;
 		case longvaluetype:
 		case ostypevaluetype:
 		case enumvaluetype:
 		case fixedvaluetype:
-			val->data.longvalue = (long) disk_to_host_int64((int64_t) disk->longvalue);
+			val->data.longvalue = disk_to_host_int64((int64_t) disk->longvalue);
 			break;
 		case singlevaluetype: {
 			double d = disk_to_host_double_bits(disk->doublebits);
